python_version/requirements-dev.txt
# Development dependencies
-r requirements.txt

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
pytest-mock==3.14.0
factory-boy==3.3.1
faker==33.3.0

# Code Quality
black==24.10.0
flake8==7.1.1
pylint==3.3.3
mypy==1.14.1
isort==5.13.2
autopep8==2.3.1

# Type Checking
django-stubs==5.1.1
djangorestframework-stubs==3.15.1
types-requests==2.32.0.20241016

# Debugging
ipdb==0.13.13
django-debug-toolbar==4.4.6
django-extensions==3.2.3

# Documentation
sphinx==8.1.3
sphinx-rtd-theme==3.0.2

# Performance
django-silk==5.3.0

# Security
bandit==1.8.0
safety==3.2.11
